Revert index and fix client transport
