+++ package.json
@@ -4,7 +4,7 @@
   "description": "anonym.club",
   "scripts": {
     "lint": "eslint .",
-    "server": "babel-node server",
+    "dev": "babel-node server/dev.server.js",
     "test": "npm run test-server; npm run test-client",
     "test-server": "mocha --compilers js:babel/register 'server/**/*.test.js'",
     "test-client": "mocha --compilers js:babel/register 'client/**/*.test.js'"
