+++ server/index.js
@@ -1,68 +1,61 @@
-/* eslint no-console: 0 */
 import express from 'express';
 import http from 'http';
-import socketIo from 'socket.io';
+import socketIO from 'socket.io';
+import _ from 'lodash';
+import * as storage from './storage';
 
-// --- SOCKET SERVER
+const app = express();
+const httpServer = new http.Server(app);
+const io = socketIO(httpServer);
 
-import routes from './routes';
+const PORT = process.env.npm_package_config_serverPort || 3001;
 
-const SOCKETHOST = process.env.SOCKETHOST || 'localhost';
-const SOCKETPORT = process.env.SOCKETPORT || 3000;
+app.use('/', express.static(__dirname + '/mock'));
 
-const socketServer = new http.Server();
-const io = socketIo(socketServer);
+io.on('connection', function onConnection(socket) {
+  socket.on('loginReq', function onLoginReq(data) {
+    const uid = _.result(data, 'uid');
 
-socketServer.listen(SOCKETPORT, () => {
-  console.log('Socket data listening on ' + SOCKETHOST + ':' + SOCKETPORT);
-});
-
-io.on('connection', (socket) => {
-  function err(error) {
-    return {
-      stat: 'ERROR',
-      data: error || 'Undefined error',
-    };
-  }
-
-  function handle(route, handler) {
-    socket.on(route, (request) => {
-      handler(request)
-        .then((res) => {
-          if (res.broadcast) {
-            io.emit('all:' + route, res.response);
-          }
-          socket.emit(route, res.response);
-        })
-        .catch((error) => {
-          socket.emit(route, err(error));
-        });
+    storage.getUser(uid).then( function onGetUser(user) {
+      if (user) {
+        socket.emit('loginRes', user);
+        io.emit('newUser', user);
+      }
+    }).catch( function createNewUser() {
+      storage.createUser().then( function onCreateUser(user) {
+        socket.emit('loginRes', user);
+        io.emit('newUser', user);
+      });
     });
-  }
+  });
 
-  Object.keys(routes)
-    .forEach((key) => handle(key, routes[key]));
+  socket.on('sendMessage', function onSendMessage(data) {
+    if (data && data.uid) {
+      storage.addUnreadMessage(data).then( function onAddUnreadMessage(message) {
+        io.emit('message', message);
+      });
+    }
+  });
+
+  socket.on('readMessage', function onReadMessage(data) {
+    storage.readMessage(_.result(data, 'mid')).then( function messageRead(message) {
+      socket.emit('messageRead', message);
+    });
+  });
+
+  socket.on('getUser', function onGetUser(data) {
+    storage.getUser(_.result(data, 'uid')).then( function sendUser(user) {
+      socket.emit('user', user);
+    });
+  });
+
+  socket.on('getRoomUsers', function onGetRoomUsers() {
+    storage.getRoomUsers().then( function sendRoomUsers(users) {
+      socket.emit('roomUsers', users);
+    });
+  });
 });
 
-// --- FILE AND HOT RELOAD SERVER
-
-import webpack from 'webpack';
-import WebpackDevServer from 'webpack-dev-server';
-import { devConfig } from '../webpack.config.babel';
-
-const FILEHOST = process.env.FILEHOST || 'localhost';
-const FILEPORT = process.env.FILEPORT || 8080;
-const FILEPATH = process.env.FILEPATH || '/../static';
-
-const fileServer = new WebpackDevServer(webpack(devConfig), {
-  publicPath: devConfig.output.publicPath,
-  hot: true,
-  historyApiFallback: true,
-  stats: {colors: true},
+httpServer.listen(PORT, function onListen() {
+  console.log('listening on *:' + PORT);
 });
-
-fileServer.use('/', express.static(__dirname + FILEPATH));
-fileServer.listen(FILEPORT, FILEHOST, () => {
-  console.log('FIle and hot reload server listening on ' + FILEHOST + ':' + FILEPORT);
-});
-
